+++ changelog.md
@@ -40,3 +40,4 @@ in certain cases
 - Improved layout of the document detail screen for larger dynamic text sizes
 - Fix for storage path not being added to newly created documents
 - Fix ASN not being saved when uploading documents with restricted user accounts
+- Reintroduce delay when typing search text